requirements.txt
flask
openai>=1.0.0
python-dotenv
twilio
requests
langdetect
backoff
